init: update recovery when enabled in settings
Update the recovery image only if the option is enabled
under Developer options

This reverts commit 231e0a9e6a1da6fa4a188840f68af649669e417f.

Change-Id: I928f7ee8bb3eaf5581bb8225661d253ecca0c4ef

Change CM recovery install script filename [2/2]

This is part 2/2 to maintain backwards compatibility with CWM's
verify_root_and_recovery() function. CWM checks if install-recovery.sh
exists and has an executable flag set, then offers to disable the script
for the user. CM now controls this with the persist.sys.recovery_update
property which is configurable via settings, so we don't need to
double-check this.

This changes the name of the recovery install script to
install-cm-recovery.sh.

Change-Id: I275dd358b46c626dfcb8fe02c583a308d5a89c56

init: Move install-recovery.sh back to the standard location

L moved the location of install-recovery.sh from /system/etc/ to
/system/bin. Since CWM recovery isn't looking for this location
anyway, let's return the file to this standard location. This allows
all other code in L to function properly.

Maintain the change to the init to allow flash_recovery to be disabled
in settings.

Change-Id: I8a85db8addeb75a2fd60d809c5ed4edc619ef7ed

Former-commit-id: 9fb68ab9698c12b19810ec954fd7b9cfbf737d4f
Former-commit-id: 2ba1c5eaa972cba446e59f7db9344a42c6573798 [formerly 2ba1c5eaa972cba446e59f7db9344a42c6573798 [formerly df2debb943a0003de4e1f1c6cd7f4f9af6aacbd2]]
Former-commit-id: 3e0b58c27735dbccc4755493dd836371d3f640df [formerly f1268213d887acc998268ff54d02cfc2c8f8d95d]
Former-commit-id: 693fa7def6250d038bc13102290cd01b7c20a331