healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I2ec9f9a420ca61814d43c316b418ce94de3691bc

Former-commit-id: dfc835ca621f3b1feb7825ca9841716ae175e1f7
Former-commit-id: 2b7212a5f095e3f95099f1c8da1f0b5e1ac66d35 [formerly 2b7212a5f095e3f95099f1c8da1f0b5e1ac66d35 [formerly 2b7212a5f095e3f95099f1c8da1f0b5e1ac66d35 [formerly bd77022fd9f147680f4c4718b5a9ed2854c4c781]]]
Former-commit-id: 23b768fb18f4495c1f4d4a306633ab0e473cbd11 [formerly 0dc48265fa3d2e8d0a8b51652276bb828bcdfaec [formerly 415e0d36ec3c732c00073366b08f094e341cee79]]
Former-commit-id: a031b6224bc36aa51273fc2315159d5ff1528f2c [formerly da98fefacaefbd74855df651871b3c276726045e]
Former-commit-id: cdfd3c184025ea461a169f4e6c9a8538422224e4