init: Weaken property override security for the init extension
Sometimes we need to override ro.* properties by using our vendor init
extension.

Previously there was a security check which was blocking that.
To resolve the issue, we need to weaken the security check during the
execution of our vendor init extension.

This is safe because the vendor init extension gets executed as part of init
construction and it is considered a trusted system component.

Change-Id: Ia7d60686968695f1fb43be4ed58770ce10da88c5

Former-commit-id: 91d12168b9500a44211a7d5a092e0ebcfed48d4d
Change-Id: I07629b6c19b5ebfa019307c497c18d8bcc719685

Former-commit-id: afc71394f9bbdd04589380bb4ef52df27f1baa26
Former-commit-id: 6bd10f8e34a25a4652b25d76326d48e38721ff5e
Signed-off-by: DennySPb <dennyspb@gmail.com>

Former-commit-id: a0bd79e9d8bd68d1f1aa88b22cd90039ab272174