ueventd: Increase uevent buffer size to 2MB
Let's increase the receive buffer size for the NETLINK_KOBJECT_UEVENT
socket to 2MB. Unless a large number of uevents queue up, that memory is
not allocated anyways. The receive buffer size serves only as an upper
limit for the total amount of memory consumed by all skbs queued to a
specific socket.

We experienced situations where ueventd got blocked for multiple seconds
while writing to /dev/kmsg, and the receive buffer overflowed in the
meantime.

Bug: 72648767
Change-Id: Ice6d7b9c5ed9c83efbad6111086ce95ac6432561

Former-commit-id: 9b5e79b26bca1a18c6e2cd22b6ac0403e2e250aa [formerly a52445e07fe3dfbf15006d725189530af1fa5433] [formerly ddf95c6ba8568e543fa1a9c4dd82cce8132ca53b [formerly bd93ad50ac64a1eaa531e5e113180720096587c3]]
Former-commit-id: ddf95c6ba8568e543fa1a9c4dd82cce8132ca53b
Former-commit-id: 7117a92c8ad09c1a50c3f7cb3033f1070335fd53