adb: make fdevent_run_on_main_thread's fd nonblocking.
If we get a ton of fdevent_run_on_main_thread calls while running one
of the handlers, the socket might become full, which will result in a
deadlock in fdevent_run_on_main_thread when a write to the fd blocks
with the mutex taken. Resolve this by making the fd nonblocking, which
is safe because we always write after appending to the list, and read
before emptying the list, which guarantees that if the byte we write is
consumed, the std::function we appended will be run.

Bug: http://b/74616284
Test: adb_test
Test: python test_device.py
Change-Id: I29319bda2ad7b5a5cdcd91d1d0ddf39f7ab7d115
(cherry picked from commit fa0ebe6b9c385698aa0dc020bef97f472772c6a2 [formerly 1222abc75b7854979e6a9f982004985179bcbede])

Former-commit-id: 424bd3588f18842f2723677e0fa31b7efd7bf478 [formerly 3b0146cc3750c51720774ff9aa7a9e398fe79faa]
Former-commit-id: 23616c544eb6d6eb4cabca3585f97ee214fdde12