init: set SetKptrRestrict before early-init
As part of debugging long run issue we are enabling
module info as part of debug build under the config key
DEBUG_MODULE_LOAD_INFO , But doing so can lead to info leak
as modules/dlkm are loaded in early-init and setting the
kptr_restriction is done after early-init.

So moving setting of kptr_restrict much before early-init.

Bug: 138641073
CRs-Fixed: 2499712

Change-Id: Ic3c0ca57a94c8b20136a15331dd646bd8825b625

Former-commit-id: 163d88aac8b2c0ee448b4b6f48b7dc77b09910bf