Make ULMK PreferredApps aware.
Add client functionality to obtain
list of preferredApps.

These apps are obtained non-frequently
at a LOW_PRESSURE event, and on a kill event
pick a non preferred app to kill instead of a
preferred app in an adj bucket.

Only if no other processes are present in an
adj bucket do we kill the preferred apps.

This ensures that user preferred apps stay longer
as cached apps and show UX concurrency benefits.

CRs-Fixed: 2522103
Change-Id: I8201d21485877c2d1af1d32c220e4b351bf236c0

Former-commit-id: 8ce685a048e29fd18a076c1890430341cce065e7