img2simg: Add support for converting holes to "don't care" chunks
This adds support for converting files with holes to "don't care"
chunks. This can result in a substantial reduction in the time it takes
to program an image if it has many holes.

Generally, constants compared to argc have been reduced by one, since we
no longer have the program name as the first argument.

Change-Id: I00750edc07d6415dcc07ae0351e9397b0222b7ba

Former-commit-id: 6150b00b6025918da8c28e5c2f929ecdf480a9d6