Revert "Add support for only starting 64 bit zygote."
This reverts commit 87f87b1f12e6a308ed79847ab603100fe993e296 [formerly da94c7f6501158de734171c0f62a486ac69ac8a9].

Reason for revert: It appears this change slows down boot on normal devices.

Technically, this change is not necessary, but it prevents starting the secondary and having it throw an error in the only run 64 bit zygote config. But it's easier to throw the error than slow down boot up.

Bug: 238971179

Test: Verified that on a 64 with 32 config, the secondary zygote
Test: starts but exits.
Change-Id: I7ab0496a402db83e70168d52e5d5911b82a3b06a

Former-commit-id: 3fa3f861d4c091aa59a6b217a3306a5ba80284f6