storaged: use correct accumulate variable for publishing disk stats
In the publish function, the variable mAccumulate_pub is used to log
disk performance. However, the variable mAccumulate is used to log disk
stats. The variable mAccumulate is only populated during a stall event.
The publish function is used to log disk stats on a periodic timer,
therefore, it should not rely on stall events to log stats. Instead,
it should rely on the appropriately named mAccumulate_pub variable.

Test: storaged unit tests
Google: 2433362
Change-Id: I30e00f2c69618705d3f308630b7e023bfc5e384e
Signed-off-by: David Munoz Constantine <dmunozc@squareup.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>

Former-commit-id: 6937dfd6a5cf5769c303cedd66af3608404b0614