snapuserd: Remove legacy xor code.
Legacy snapuserd had xor optimizations implemented since at the time the
userspace snapshot implementation was a work in progress. We forgot to
ultimately remove the legacy xor code however, so let's do that now.

This should have no impact on devices. An S-frozen device will not be
getting this code because there's no vendor update happening. A T-frozen
device will be using userspace snapshots.

The only scenario this could happen is if a device starts on S vendor,
upgrades to T vendor, enables xor compression, and then updates to U
vendor. This configuration is not supported. Any device with a T+ vendor
should be using userspace snapshots.

Bug: N/A
Test: vabc_legacy_tests
Change-Id: If484a53cb0527393898a7ea703ba6f12d4c9a310

Former-commit-id: 0172b1cc34ba49d74ac02bf5e5e38fe7a818603a