libfiemap: Handle EAGAIN in fallocate().
When we changed our ENOSPC tests, it exposed a path in F2FS were
fallocate can return EAGAIN. This is expected if F2FS attempts GC to
acquire more chunks, and it can leave the file in a partially allocated
state.

As a fix, keep attempting fallocate() as long as (1) it returns EAGAIN,
and (2) the allocated size keeps growing. If (2) fails we return ENOSPC.

Bug: N/A
Test: treehugger
Google: 2590526
Change-Id: I5f867b5a200b9260e486985f203f9872a949b3f9
Signed-off-by: DennySPb <dennyspb@gmail.com>

Former-commit-id: 93688acbb1c96755cd35ec4fb4be9b8b7d563472