Move ENOSPC tests to libfiemap.
These tests are still giving us trouble. Move them to libfiemap, which
(1) is closer to the source of implementation, and (2) allows us to
re-use the temporary filesystem code. This won't perturb the state of
the actual device.

The new test creates a 64MB ext4 or f2fs mount point as a sandbox, which
should be much safer.

Bug: 285197715
Bug: 298346574
Bug: 299142557
Test: fiemap_writer_test
Merged-In: I33502d49613be4f269a80e5c632514fc56a0246a
Ignore-AOSP-First: cherry-pick
Change-Id: Iedb7c32a594c3b1fca2904f3441029aaed7edf2a

Former-commit-id: 3a60e825126bbcfebecd57a728a1a87cacc5b497